+++ src/DataFixtures/ArticleFixtures.php
@@ -5,11 +5,26 @@ namespace App\DataFixtures;
 use App\Entity\Article;
 use App\Entity\Comment;
 use App\Entity\Tag;
+use App\Service\UploaderHelper;
 use Doctrine\Common\DataFixtures\DependentFixtureInterface;
 use Doctrine\Common\Persistence\ObjectManager;
+use Symfony\Component\Filesystem\Filesystem;
+use Symfony\Component\HttpFoundation\File\File;
 
 class ArticleFixtures extends BaseFixture implements DependentFixtureInterface
 {
+    private $uploaderHelper;
+
+    /**
+     * We want to use it to fake some files in fixtures
+     * ArticleFixtures constructor.
+     * @param UploaderHelper $uploaderHelper
+     */
+    public function __construct(UploaderHelper $uploaderHelper)
+    {
+        $this->uploaderHelper = $uploaderHelper;
+    }
+
     private static $articleTitles = [
         'Why Asteroids Taste Like Bacon',
         'Life on Planet Mercury: Tan, Relaxing and Fabulous',
@@ -52,9 +67,10 @@ EOF
                 $article->setPublishedAt($this->faker->dateTimeBetween('-100 days', '-1 days'));
             }
 
+            $imageFilename = $this->fakeUploadImage();
             $article->setAuthor($this->getRandomReference('main_users'))
                 ->setHeartCount($this->faker->numberBetween(5, 100))
-                ->setImageFilename($this->faker->randomElement(self::$articleImages))
+                ->setImageFilename($imageFilename)
             ;
 
             $tags = $this->getRandomReferences('main_tags', $this->faker->numberBetween(0, 5));
@@ -66,6 +82,19 @@ EOF
         });
 
         $manager->flush();
+    }
+
+    /**
+     * We want to pretend that we already uploaded these images by uploader but with coping not moving it
+     */
+    private function fakeUploadImage(): string
+    {
+        $randomImage = $this->faker->randomElement(self::$articleImages);
+        $fs = new Filesystem();
+        $targetPath = sys_get_temp_dir().'/'.$randomImage;
+        $fs->copy(__DIR__.'/images/'.$randomImage, $targetPath, true);
+        return $this->uploaderHelper
+            ->uploadArticleImage(new File($targetPath));
     }
 
     public function getDependencies()